# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PartyNest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /availability/{businessId}:
    get:
      operationId: getAvailability
      tags: [availability]
      summary: Get availability slots for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"

    post:
      operationId: setAvailability
      tags: [availability]
      summary: Set availability slot (business owner)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAvailabilityRequest"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (for current user or business)
      parameters:
        - name: businessId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (accept/reject/cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusRequest"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /crm/dashboard:
    get:
      operationId: getCrmDashboard
      tags: [crm]
      summary: Get CRM dashboard stats for a business
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmDashboard"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [parent, business, admin]
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, createdAt]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [parent, business]
      required: [email, password, name, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
    Business:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        locationType:
          type: string
          enum: [indoor, outdoor, both]
        minAge:
          type: integer
        maxAge:
          type: integer
        capacity:
          type: integer
        basePrice:
          type: number
        photos:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        rules:
          type: string
        phone:
          type: string
        website:
          type: string
        visibilityPoints:
          type: integer
        views:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, name, city, address, locationType, capacity, basePrice, createdAt]
...
    Package:
      type: object
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        priceWeekend:
          type: number
        maxGuests:
          type: integer
        duration:
          type: integer
        includes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [id, businessId, name, price, maxGuests, duration]
...
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: string
        businessId:
          type: string
        date:
          type: string
          format: date
        timeSlot:
          type: string
          enum: [morning, afternoon, evening]
        isAvailable:
          type: boolean
        isBlocked:
          type: boolean
        note:
          type: string
      required: [id, businessId, date, timeSlot, isAvailable]
    SetAvailabilityRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        timeSlot:
          type: string
          enum: [morning, afternoon, evening]
        isAvailable:
          type: boolean
        isBlocked:
          type: boolean
        note:
          type: string
      required: [date, timeSlot]
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessId:
          type: string
        packageId:
          type: string
        date:
          type: string
          format: date
        timeSlot:
          type: string
          enum: [morning, afternoon, evening]
        guestCount:
          type: integer
        childrenAge:
          type: string
        notes:
          type: string
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
        totalPrice:
          type: number
        business:
          $ref: "#/components/schemas/Business"
        package:
          $ref: "#/components/schemas/Package"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, businessId, date, timeSlot, guestCount, parentName, parentEmail, parentPhone, status, createdAt]
    CreateBookingRequest:
      type: object
      properties:
        businessId:
          type: string
        packageId:
          type: string
        date:
          type: string
          format: date
        timeSlot:
          type: string
          enum: [morning, afternoon, evening]
        guestCount:
          type: integer
        childrenAge:
          type: string
        notes:
          type: string
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        totalPrice:
          type: number
      required: [businessId, date, timeSlot, guestCount, parentName, parentEmail, parentPhone]
    UpdateBookingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, cancelled]
        note:
          type: string
      required: [status]
    Quote:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
        childrenAge:
          type: string
        message:
          type: string
        services:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, responded, closed]
        business:
          $ref: "#/components/schemas/Business"
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, name, email, guestCount, status, createdAt]
...
    CrmDashboard:
      type: object
      properties:
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        acceptedBookings:
          type: integer
        totalViews:
          type: integer
        totalQuotes:
          type: integer
        visibilityPoints:
          type: integer
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        recentQuotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        monthlyStats:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyStat"
      required: [totalBookings, pendingBookings, acceptedBookings, totalViews, totalQuotes, visibilityPoints, recentBookings, recentQuotes, monthlyStats]
    MonthlyStat:
      type: object
      properties:
        month:
          type: string
        bookings:
          type: integer
        views:
          type: integer
      required: [month, bookings, views]
...