# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PartyNest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilitySlotTimeSlot } from "./availabilitySlotTimeSlot";

export interface AvailabilitySlot {
  id: string;
  businessId: string;
  date: Date;
  timeSlot: AvailabilitySlotTimeSlot;
  isAvailable: boolean;
  isBlocked?: boolean;
  note?: string;
}