# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PartyNest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestTimeSlot } from "./createBookingRequestTimeSlot";

export interface CreateBookingRequest {
  businessId: string;
  packageId?: string;
  date: Date;
  timeSlot: CreateBookingRequestTimeSlot;
  guestCount: number;
  childrenAge?: string;
  notes?: string;
  parentName: string;
  parentEmail: string;
  parentPhone: string;
  totalPrice?: number;
}